Parser1/requirements.txt
beautifulsoup4==4.12.2
fastapi==0.100.0
pandas==2.0.3
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0
urllib3==2.0.3
uvicorn==0.22.0
pytest==7.4.0
python-docx==0.8.11
